fix 모든 레거시 코드 수정, 데이터 테이블 관련 내용 전체 정리, ID 항목 정리, 대미지 계산 및 에이전트 스킬 관련 내용 정리

+++ Assets/Project/Scripts/Systems/DamageCalculator.cs
@@ -9,7 +9,7 @@ public static class DamageCalculator
     /// 공격자의 ATK, 방어자의 DEF, 크리티컬 확률/배수로 최종 피해량 계산
     /// isCrit: 이번 공격이 크리티컬인지 out 반환
     /// </summary>
-    public static int ComputeCharacterDamage(int atk, int def, float critChance, float critMultiplier, out bool isCrit)
+    public static int ComputeCharacterDamage(float atk, float def, float critChance, float critMultiplier, out bool isCrit)
     {
         float baseDmg = Mathf.Max(1f, atk - def);
         isCrit = Random.value < critChance;
@@ -20,7 +20,7 @@ public static class DamageCalculator
     /// <summary>
     /// isCrit 없이 사용하는 오버로드 (하위 호환)
     /// </summary>
-    public static int ComputeCharacterDamage(int atk, int def, float critChance, float critMultiplier)
+    public static int ComputeCharacterDamage(float atk, float def, float critChance, float critMultiplier)
     {
         bool _ ;
         return ComputeCharacterDamage(atk, def, critChance, critMultiplier, out _);
